+++ src/cgame/cg_window.c
@@ -41,6 +41,9 @@ extern pmove_t cg_pmove;        // cg_predict.c
 
 vec4_t colorGreen2 = { 0.305f, 0.475f, 0.305f, 0.48f }; // Slightly off from default fill
 
+/**
+ * @brief CG_createStatsWindow
+ */
 void CG_createStatsWindow(void)
 {
 	cg_window_t *sw = CG_windowAlloc(WFX_TEXTSIZING | WFX_FADEIN | /*WFX_SCROLLUP|*/ WFX_TRUETYPE, 110);
@@ -59,6 +62,9 @@ void CG_createStatsWindow(void)
 	sw->y          = (cg.snap->ps.pm_type == PM_INTERMISSION) ? -20 : -100; // Align from bottom minus offset and height
 }
 
+/**
+ * @brief CG_createTopShotsWindow
+ */
 void CG_createTopShotsWindow(void)
 {
 	cg_window_t *sw = CG_windowAlloc(WFX_TEXTSIZING | WFX_FLASH | WFX_FADEIN | WFX_SCROLLUP | WFX_TRUETYPE, 190);
@@ -79,6 +85,9 @@ void CG_createTopShotsWindow(void)
 	memcpy(&sw->colorBackground2, &colorGreen2, sizeof(vec4_t));
 }
 
+/**
+ * @brief CG_createMOTDWindow
+ */
 void CG_createMOTDWindow(void)
 {
 	const char *str = CG_ConfigString(CS_CUSTMOTD + 0);
@@ -128,7 +137,10 @@ void CG_createMOTDWindow(void)
 //////////////////////////////////////////////
 //////////////////////////////////////////////
 
-// Windowing system setup
+
+/**
+ * @brief Windowing system setup
+ */
 void CG_windowInit(void)
 {
 	int i;
@@ -145,7 +157,12 @@ void CG_windowInit(void)
 	cg.topshotsWindow     = NULL;
 }
 
-// Window stuct "constructor" with some common defaults
+/**
+ * @brief Window stuct "constructor" with some common defaults
+ * @param[in] w
+ * @param[out] fx
+ * @param[out] startupLength
+ */
 void CG_windowReset(cg_window_t *w, int fx, int startupLength)
 {
 	vec4_t colorGeneralBorder = { 0.5f, 0.35f, 0.25f, 0.5f };
@@ -169,7 +186,12 @@ void CG_windowReset(cg_window_t *w, int fx, int startupLength)
 	memcpy(&w->colorBackground, &colorGeneralFill, sizeof(vec4_t));
 }
 
-// Reserve a window
+/**
+ * @brief Reserve a window
+ * @param[in] fx
+ * @param[in] startupLength
+ * @return 
+ */
 cg_window_t *CG_windowAlloc(int fx, int startupLength)
 {
 	int                i;
@@ -196,7 +218,10 @@ cg_window_t *CG_windowAlloc(int fx, int startupLength)
 	return NULL;
 }
 
-// Free up a window reservation
+/**
+ * @brief Free up a window reservation
+ * @param[in,out] w
+ */
 void CG_windowFree(cg_window_t *w)
 {
 	int                i, j;
@@ -239,6 +264,9 @@ void CG_windowFree(cg_window_t *w)
 	}
 }
 
+/**
+ * @brief CG_windowCleanup
+ */
 void CG_windowCleanup(void)
 {
 	int                i;
@@ -256,6 +284,9 @@ void CG_windowCleanup(void)
 	}
 }
 
+/**
+ * @brief CG_demoAviFPSDraw
+ */
 void CG_demoAviFPSDraw(void)
 {
 	qboolean fKeyDown = (qboolean)(cgs.fKeyPressed[K_F1] | cgs.fKeyPressed[K_F2] | cgs.fKeyPressed[K_F3] | cgs.fKeyPressed[K_F4] | cgs.fKeyPressed[K_F5]);
@@ -266,6 +297,9 @@ void CG_demoAviFPSDraw(void)
 	}
 }
 
+/**
+ * @brief CG_demoTimescaleDraw
+ */
 void CG_demoTimescaleDraw(void)
 {
 	if (cg.demoPlayback && cgs.timescaleUpdate > cg.time && demo_drawTimeScale.integer != 0)
@@ -273,18 +307,20 @@ void CG_demoTimescaleDraw(void)
 		vec4_t bgColor = { 0.0f, 0.0f, 0.0f, 0.6f };
 		vec4_t bdColor = { 0.5f, 0.5f, 0.5f, 0.5f };
 
-		char *s = va("^7Time Scale: ^3%.1fx", cg_timescale.value);
+		char *s = va("^7Time Scale: ^3%.1fx", (double)cg_timescale.value);
 		int  h  = CG_Text_Height_Ext("A", cg_fontScaleSP.value, 0, &cgs.media.limboFont2);
 		int  w  = CG_Text_Width_Ext(s, cg_fontScaleSP.value, 0, &cgs.media.limboFont2);
-		int  x  = Ccg_WideX(SCREEN_WIDTH) - w - 108;
+		int  x  = (int)Ccg_WideX(SCREEN_WIDTH) - w - 108;
 
 		CG_FillRect(x, SCREEN_HEIGHT - 21, w + 7, h * 2.5f, bgColor);
 		CG_DrawRect(x, SCREEN_HEIGHT - 21, w + 7, h * 2.5f, 1, bdColor);
 		CG_Text_Paint_Ext(x + 3, SCREEN_HEIGHT - 10, cg_fontScaleSP.value, cg_fontScaleSP.value, colorWhite, s, 0, 0, 0, &cgs.media.limboFont2);
 	}
 }
-
-// Main window-drawing handler
+ 
+/**
+ * @brief Main window-drawing handler
+ */
 void CG_windowDraw(void)
 {
 	int         h, x, y, i, j, milli, t_offset, tmp;
@@ -346,9 +382,9 @@ void CG_windowDraw(void)
 
 		bg = ((w->effects & WFX_FLASH) && (milli % w->flashPeriod) > w->flashMidpoint) ? &w->colorBackground2 : &w->colorBackground;
 
-		h            = w->h;
-		x            = w->x;
-		y            = w->y;
+		h            = (int)w->h;
+		x            = (int)w->x;
+		y            = (int)w->y;
 		t_offset     = milli - w->time;
 		textColor[3] = 1.0f;
 		memcpy(&borderColor, w->colorBorder, sizeof(vec4_t));
@@ -390,7 +426,7 @@ void CG_windowDraw(void)
 			{
 				if (tmp > 0)
 				{
-					y = w->curY + (SCREEN_HEIGHT - w->y) * t_offset / w->targetTime;        //(100 * t_offset / w->targetTime) / 100;
+					y = (int)(w->curY + (SCREEN_HEIGHT - w->y) * t_offset / w->targetTime);        //(100 * t_offset / w->targetTime) / 100;
 				}
 				if (tmp < 0 || y >= SCREEN_HEIGHT)
 				{
@@ -464,7 +500,10 @@ void CG_windowDraw(void)
 	}
 }
 
-// Set the window width and height based on the windows text/font parameters
+/**
+ * @brief Set the window width and height based on the windows text/font parameters
+ * @param[in,out] w
+ */
 void CG_windowNormalizeOnText(cg_window_t *w)
 {
 	int i, tmp;
@@ -479,8 +518,8 @@ void CG_windowNormalizeOnText(cg_window_t *w)
 
 	if (!(w->effects & WFX_TRUETYPE))
 	{
-		w->fontWidth  = w->fontScaleX * WINDOW_FONTWIDTH;
-		w->fontHeight = w->fontScaleY * WINDOW_FONTHEIGHT;
+		w->fontWidth  = (int)(w->fontScaleX * WINDOW_FONTWIDTH);
+		w->fontHeight = (int)(w->fontScaleY * WINDOW_FONTHEIGHT);
 	}
 
 	for (i = 0; i < w->lineCount; i++)
@@ -529,6 +568,10 @@ void CG_windowNormalizeOnText(cg_window_t *w)
 	}
 }
 
+/**
+ * @brief CG_printWindow
+ * @param[in] str
+ */
 void CG_printWindow(const char *str)
 {
 	int         pos = 0, pos2 = 0;
@@ -572,8 +615,10 @@ void CG_printWindow(const char *str)
 		CG_addString(w, buf + pos2);
 	}
 }
-
-// String buffer handling
+ 
+/**
+ * @brief String buffer handling
+ */
 void CG_initStrings(void)
 {
 	int i;
@@ -585,6 +630,12 @@ void CG_initStrings(void)
 	}
 }
 
+/**
+ * @brief CG_addString
+ * @param[in,out] w
+ * @param[in] buf
+ * @return 
+ */
 qboolean CG_addString(cg_window_t *w, const char *buf)
 {
 	int i;
@@ -625,6 +676,10 @@ qboolean CG_addString(cg_window_t *w, const char *buf)
 	return qfalse;
 }
 
+/**
+ * @brief CG_removeStrings
+ * @param[in,out] w
+ */
 void CG_removeStrings(cg_window_t *w)
 {
 	int i, j;
@@ -655,7 +710,9 @@ void CG_removeStrings(cg_window_t *w)
 // cgame cursor handling
 
 #ifdef FEATURE_MULTIVIEW
-// Mouse overlay for controlling multiview windows
+/**
+ * @brief Mouse overlay for controlling multiview windows
+ */
 void CG_cursorUpdate(void)
 {
 	int                i, x;
@@ -806,7 +863,7 @@ void CG_cursorUpdate(void)
 	{
 		// display on two columns
 		int hOffset = 32;
-		int xOffset = MVINFO_RIGHT - hOffset;
+		int xOffset = (int)MVINFO_RIGHT - hOffset;
 
 		// Ugh, have to loop through BOTH team lists
 		for (i = TEAM_AXIS; i <= TEAM_ALLIES; i++)
@@ -819,12 +876,12 @@ void CG_cursorUpdate(void)
 			if (cgs.cursorX >= nx && cgs.cursorY >= ny && cgs.cursorX < xOffset &&
 			    cgs.cursorY < ny + (cg.mvTotalTeam[i] * charHeight * 2.0f))
 			{
-				int pos = (int)(cgs.cursorY - ny) / (charHeight * 2.0f);
+				int pos = (int)((cgs.cursorY - ny) / (charHeight * 2.0f));
 
 				if (pos < cg.mvTotalTeam[i])
 				{
 					int x = xOffset - cg.mvOverlay[(cg.mvTeamList[i][pos])].width;
-					int y = MVINFO_TOP + ((pos + 1) * charHeight * 2.0f);
+					int y = (int)(MVINFO_TOP + ((pos + 1) * charHeight * 2.0f));
 
 					// See if we're really over something
 					if (cgs.cursorX >= x && cgs.cursorY >= y &&
